+++ __tests__/fixtures/browser.js
@@ -3,11 +3,8 @@
 function modifyText() {
   const t2 = document.querySelector("#t2");
 
-  if (t2.firstChild.nodeValue === "three") {
-    t2.firstChild.nodeValue = "two";
-  } else {
-    t2.firstChild.nodeValue = "three";
-  }
+  t2.firstChild.nodeValue =
+    t2.firstChild.nodeValue === "three" ? "two" : "three";
 }
 
 // Add event listener to table